impr: Refactoring of floating point tool to decrease code complexity. (#1275)
I noticed the bad score on code factor so I reorganized it to make it
more readable and maintainable. In order to break down the big function
into it much smaller parts I encapsulated all the variables that the
functions need to access in two classes, one for the imgui related
statics and the other for non-static variables.

When writing the smaller functions I was noticed that there was room to
simplify the existing algorithms by writing functions that could be
called by parts that previously shared no code. I tested the changes the
same way I tested the original and it seems to work the same way but
maybe a bit faster. Although it may be possible to further optimize the
present code code factor no longer flags the function at all.

+++ plugins/builtin/source/content/data_processor_nodes.cpp
@@ -563,7 +563,7 @@ namespace hex::plugin::builtin {
         void process() override {
             const auto &input = this->getBufferOnInput(0);
 
-            float output = 0;
+            double output = 0;
             if (input.empty() || input.size() != sizeof(output))
                 throwNodeError("Buffer is empty or not the right size to fit a float");
